+++ src/components/DarkModeToggle.tsx
@@ -5,29 +5,49 @@ import { ToolButton } from "./ToolButton";
 import { THEME } from "../constants";
 import { Theme } from "../element/types";
 
-// We chose to use only explicit toggle and not a third option for system value,
-// but this could be added in the future.
+// Historically, this toggle was not supporting a system value, however
+// now it might be renamed to something more appropriate.
 export const DarkModeToggle = (props: {
   value: Theme;
   onChange: (value: Theme) => void;
   title?: string;
+  withSystemTheme?: boolean;
 }) => {
-  const title =
-    props.title ||
-    (props.value === "dark" ? t("buttons.lightMode") : t("buttons.darkMode"));
+  let title = props.title || t("buttons.lightMode");
+  if (!props.title) {
+    switch (props.value) {
+      case THEME.DARK:
+        title = t("buttons.darkMode");
+        break;
+      case THEME.SYSTEM:
+        t("buttons.autoMode");
+        break;
+    }
+  }
 
   return (
     <ToolButton
       type="icon"
-      icon={props.value === THEME.LIGHT ? ICONS.MOON : ICONS.SUN}
+      icon={THEME_TO_ICON[props.value] ?? THEME.LIGHT}
       title={title}
       aria-label={title}
       onClick={() =>
-        props.onChange(props.value === THEME.DARK ? THEME.LIGHT : THEME.DARK)
+        props.onChange(getNextTheme(props.value, props.withSystemTheme))
       }
       data-testid="toggle-dark-mode"
     />
   );
+};
+
+const getNextTheme = (previous: Theme, withSystemTheme = false) => {
+  switch (previous) {
+    case THEME.DARK:
+      return withSystemTheme ? THEME.SYSTEM : THEME.LIGHT;
+    case THEME.SYSTEM:
+      return THEME.LIGHT;
+    default:
+      return THEME.DARK;
+  }
 };
 
 const ICONS = {
@@ -47,4 +67,18 @@ const ICONS = {
       ></path>
     </svg>
   ),
+  MOON_AND_SUN: (
+    <svg width="512" height="512" className="rtl-mirror" viewBox="0 0 512 512">
+      <path
+        fill="currentColor"
+        d="M230 506.465C114.09 481.972 27.211 379.084 27.211 256c0-123.222 86.989-225.98 202.79-250.46v500.925Zm50-23.285V381.479c24.017-4.748 46.927-16.406 65.5-34.979 49.9-49.9 49.9-131.1 0-181-18.573-18.573-41.483-30.231-65.5-34.979V28.6l37.9 75.8 100.4-33.5c13.5-4.6 26.4 8.3 21.9 21.9l-33.5 100.4 94.7 47.3c12.8 6.4 12.8 24.6-.2 31.1l-94.7 47.3L440 419.3c4.6 13.5-8.3 26.4-21.9 21.9l-100.4-33.5-37.7 75.48Zm0-134.489V163.31c40.833 11.04 71 48.435 71 92.691s-30.167 81.652-71 92.691Z"
+      />
+    </svg>
+  ),
+};
+
+const THEME_TO_ICON = {
+  [THEME.LIGHT]: ICONS.SUN,
+  [THEME.DARK]: ICONS.MOON,
+  [THEME.SYSTEM]: ICONS.MOON_AND_SUN,
 };